Move `IpDisplayBuffer` into submodule.
